chore(api): move helpers into `internal/api`
