+++ api/v1alpha1/event.go
@@ -1,13 +1,5 @@
 package v1alpha1
 
-import (
-	"time"
-
-	authnv1 "k8s.io/api/authentication/v1"
-
-	"github.com/akuity/kargo/internal/server/user"
-)
-
 const (
 	AnnotationKeyEventActor                  = "event.kargo.akuity.io/actor"
 	AnnotationKeyEventProject                = "event.kargo.akuity.io/project"
@@ -50,57 +42,3 @@ const (
 	EventActorKubernetesUserPrefix = "kubernetes:"
 	EventActorUnknown              = "unknown actor"
 )
-
-func FormatEventControllerActor(name string) string {
-	return EventActorControllerPrefix + name
-}
-
-// FormatEventUserActor returns a string representation of the user acting in an event
-// that can be used as a value of AnnotationKeyEventActor.
-//
-// 1. If the user is admin, it returns EventActorAdmin since other information is not available.
-// 2. If the email is available, it returns email in "email:<email>" format.
-// 3. If the subject is available, it returns subject in "subject:<subject>" format.
-// 4. Otherwise, it returns EventActorUnknown.
-func FormatEventUserActor(u user.Info) string {
-	var email, subject string
-	if emailClaim, ok := u.Claims["email"]; ok {
-		if emailStr, ok := emailClaim.(string); ok {
-			email = emailStr
-		}
-	}
-	if subClaim, ok := u.Claims["sub"]; ok {
-		if subStr, ok := subClaim.(string); ok {
-			subject = subStr
-		}
-	}
-
-	switch {
-	case u.IsAdmin:
-		return EventActorAdmin
-	case email != "":
-		return EventActorEmailPrefix + email
-	case subject != "":
-		return EventActorSubjectPrefix + subject
-	default:
-		return EventActorUnknown
-	}
-}
-
-func FormatEventKubernetesUserActor(u authnv1.UserInfo) string {
-	return EventActorKubernetesUserPrefix + u.Username
-}
-
-func NewFreightApprovedEventAnnotations(actor string, f *Freight, stageName string) map[string]string {
-	annotations := map[string]string{
-		AnnotationKeyEventProject:           f.Namespace,
-		AnnotationKeyEventFreightCreateTime: f.CreationTimestamp.Format(time.RFC3339),
-		AnnotationKeyEventFreightAlias:      f.Alias,
-		AnnotationKeyEventFreightName:       f.Name,
-		AnnotationKeyEventStageName:         stageName,
-	}
-	if actor != "" {
-		annotations[AnnotationKeyEventActor] = actor
-	}
-	return annotations
-}
