+++ internal/api/annotations_test.go
@@ -1,16 +1,18 @@
-package v1alpha1
+package api
 
 import (
 	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/require"
+
+	kargoapi "github.com/akuity/kargo/api/v1alpha1"
 )
 
 func TestRefreshAnnotationValue(t *testing.T) {
 	t.Run("has refresh annotation", func(t *testing.T) {
 		result, ok := RefreshAnnotationValue(map[string]string{
-			AnnotationKeyRefresh: "foo",
+			kargoapi.AnnotationKeyRefresh: "foo",
 		})
 		require.True(t, ok)
 		require.Equal(t, "foo", result)
@@ -24,7 +26,7 @@ func TestRefreshAnnotationValue(t *testing.T) {
 
 	t.Run("has refresh annotation with empty value", func(t *testing.T) {
 		result, ok := RefreshAnnotationValue(map[string]string{
-			AnnotationKeyRefresh: "",
+			kargoapi.AnnotationKeyRefresh: "",
 		})
 		require.True(t, ok)
 		require.Empty(t, result)
@@ -34,7 +36,7 @@ func TestRefreshAnnotationValue(t *testing.T) {
 func TestReverifyAnnotationValue(t *testing.T) {
 	t.Run("has reverify annotation with valid JSON", func(t *testing.T) {
 		result, ok := ReverifyAnnotationValue(map[string]string{
-			AnnotationKeyReverify: `{"id":"foo"}`,
+			kargoapi.AnnotationKeyReverify: `{"id":"foo"}`,
 		})
 		require.True(t, ok)
 		require.Equal(t, "foo", result.ID)
@@ -42,7 +44,7 @@ func TestReverifyAnnotationValue(t *testing.T) {
 
 	t.Run("has reverify annotation with ID string", func(t *testing.T) {
 		result, ok := ReverifyAnnotationValue(map[string]string{
-			AnnotationKeyReverify: "foo",
+			kargoapi.AnnotationKeyReverify: "foo",
 		})
 		require.True(t, ok)
 		require.Equal(t, "foo", result.ID)
@@ -56,7 +58,7 @@ func TestReverifyAnnotationValue(t *testing.T) {
 
 	t.Run("has reverify annotation with empty ID", func(t *testing.T) {
 		result, ok := ReverifyAnnotationValue(map[string]string{
-			AnnotationKeyAbort: "",
+			kargoapi.AnnotationKeyAbort: "",
 		})
 		require.False(t, ok)
 		require.Nil(t, result)
@@ -66,7 +68,7 @@ func TestReverifyAnnotationValue(t *testing.T) {
 func TestAbortVerificationAnnotationValue(t *testing.T) {
 	t.Run("has abort annotation with valid JSON", func(t *testing.T) {
 		result, ok := AbortVerificationAnnotationValue(map[string]string{
-			AnnotationKeyAbort: `{"id":"foo"}`,
+			kargoapi.AnnotationKeyAbort: `{"id":"foo"}`,
 		})
 		require.True(t, ok)
 		require.Equal(t, "foo", result.ID)
@@ -74,7 +76,7 @@ func TestAbortVerificationAnnotationValue(t *testing.T) {
 
 	t.Run("has abort annotation with ID string", func(t *testing.T) {
 		result, ok := AbortVerificationAnnotationValue(map[string]string{
-			AnnotationKeyAbort: "foo",
+			kargoapi.AnnotationKeyAbort: "foo",
 		})
 		require.True(t, ok)
 		require.Equal(t, "foo", result.ID)
@@ -88,7 +90,7 @@ func TestAbortVerificationAnnotationValue(t *testing.T) {
 
 	t.Run("has abort annotation with empty ID", func(t *testing.T) {
 		result, ok := AbortVerificationAnnotationValue(map[string]string{
-			AnnotationKeyAbort: "",
+			kargoapi.AnnotationKeyAbort: "",
 		})
 		require.False(t, ok)
 		require.Nil(t, result)
@@ -98,18 +100,18 @@ func TestAbortVerificationAnnotationValue(t *testing.T) {
 func TestAbortPromotionAnnotationValue(t *testing.T) {
 	t.Run("has abort annotation with valid JSON", func(t *testing.T) {
 		result, ok := AbortPromotionAnnotationValue(map[string]string{
-			AnnotationKeyAbort: fmt.Sprintf(`{"action":"%s"}`, AbortActionTerminate),
+			kargoapi.AnnotationKeyAbort: fmt.Sprintf(`{"action":"%s"}`, kargoapi.AbortActionTerminate),
 		})
 		require.True(t, ok)
-		require.Equal(t, AbortActionTerminate, result.Action)
+		require.Equal(t, kargoapi.AbortActionTerminate, result.Action)
 	})
 
 	t.Run("has abort annotation with action string", func(t *testing.T) {
 		result, ok := AbortPromotionAnnotationValue(map[string]string{
-			AnnotationKeyAbort: string(AbortActionTerminate),
+			kargoapi.AnnotationKeyAbort: string(kargoapi.AbortActionTerminate),
 		})
 		require.True(t, ok)
-		require.Equal(t, AbortActionTerminate, result.Action)
+		require.Equal(t, kargoapi.AbortActionTerminate, result.Action)
 	})
 
 	t.Run("does not have abort annotation", func(t *testing.T) {
@@ -120,7 +122,7 @@ func TestAbortPromotionAnnotationValue(t *testing.T) {
 
 	t.Run("has abort annotation with empty action", func(t *testing.T) {
 		result, ok := AbortPromotionAnnotationValue(map[string]string{
-			AnnotationKeyAbort: "",
+			kargoapi.AnnotationKeyAbort: "",
 		})
 		require.False(t, ok)
 		require.Nil(t, result)
