+++ internal/api/freight.go
@@ -1,0 +1,112 @@
+package api
+
+import (
+	"context"
+	"fmt"
+
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	kargoapi "github.com/akuity/kargo/api/v1alpha1"
+)
+
+// GetFreightByNameOrAlias returns a pointer to the Freight resource specified
+// by the project, and name OR alias arguments. If no such resource is found,
+// nil is returned instead.
+func GetFreightByNameOrAlias(
+	ctx context.Context,
+	c client.Client,
+	project string,
+	name string,
+	alias string,
+) (*kargoapi.Freight, error) {
+	if name != "" {
+		return GetFreight(
+			ctx,
+			c,
+			types.NamespacedName{
+				Namespace: project,
+				Name:      name,
+			},
+		)
+	}
+	return GetFreightByAlias(ctx, c, project, alias)
+}
+
+// GetFreight returns a pointer to the Freight resource specified by the
+// namespacedName argument. If no such resource is found, nil is returned
+// instead.
+func GetFreight(
+	ctx context.Context,
+	c client.Client,
+	namespacedName types.NamespacedName,
+) (*kargoapi.Freight, error) {
+	freight := kargoapi.Freight{}
+	if err := c.Get(ctx, namespacedName, &freight); err != nil {
+		if err = client.IgnoreNotFound(err); err == nil {
+			return nil, nil
+		}
+		return nil, fmt.Errorf(
+			"error getting Freight %q in namespace %q: %w",
+			namespacedName.Name,
+			namespacedName.Namespace,
+			err,
+		)
+	}
+	return &freight, nil
+}
+
+// GetFreightByAlias returns a pointer to the Freight resource specified by the
+// project and alias arguments. If no such resource is found, nil is returned
+// instead.
+func GetFreightByAlias(
+	ctx context.Context,
+	c client.Client,
+	project string,
+	alias string,
+) (*kargoapi.Freight, error) {
+	freightList := kargoapi.FreightList{}
+	if err := c.List(
+		ctx,
+		&freightList,
+		client.InNamespace(project),
+		client.MatchingLabels{
+			kargoapi.AliasLabelKey: alias,
+		},
+	); err != nil {
+		return nil, fmt.Errorf(
+			"error listing Freight with alias %q in namespace %q: %w",
+			alias,
+			project,
+			err,
+		)
+	}
+	if len(freightList.Items) == 0 {
+		return nil, nil
+	}
+	return &freightList.Items[0], nil
+}
+
+// ListFreightByCurrentStage returns a list of Freight resources that think
+// they're currently in use by the Stage specified.
+func ListFreightByCurrentStage(
+	ctx context.Context,
+	c client.Client,
+	stage *kargoapi.Stage,
+) ([]kargoapi.Freight, error) {
+	freightList := kargoapi.FreightList{}
+	if err := c.List(
+		ctx,
+		&freightList,
+		client.InNamespace(stage.Namespace),
+		client.MatchingFields{"currentlyIn": stage.Name},
+	); err != nil {
+		return nil, fmt.Errorf(
+			"error listing Freight in namespace %q with current stage %q: %w",
+			stage.Namespace,
+			stage.Name,
+			err,
+		)
+	}
+	return freightList.Items, nil
+}
