+++ internal/api/warehouse_test.go
@@ -1,4 +1,4 @@
-package v1alpha1
+package api
 
 import (
 	"context"
@@ -14,6 +14,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	"sigs.k8s.io/controller-runtime/pkg/client/interceptor"
+
+	kargoapi "github.com/akuity/kargo/api/v1alpha1"
 )
 
 // TODO(krancour): If we move our actual indexers to this package, we can use
@@ -22,13 +24,13 @@ import (
 const warehouseField = "warehouse"
 
 func warehouseIndexer(obj client.Object) []string {
-	return []string{obj.(*Freight).Origin.Name} // nolint: forcetypeassert
+	return []string{obj.(*kargoapi.Freight).Origin.Name} // nolint: forcetypeassert
 }
 
 const approvedField = "approvedFor"
 
 func approvedForIndexer(obj client.Object) []string {
-	freight := obj.(*Freight) // nolint: forcetypeassert
+	freight := obj.(*kargoapi.Freight) // nolint: forcetypeassert
 	var approvedFor []string
 	for stage := range freight.Status.ApprovedFor {
 		approvedFor = append(approvedFor, stage)
@@ -39,7 +41,7 @@ func approvedForIndexer(obj client.Object) []string {
 const verifiedInField = "verifiedIn"
 
 func verifiedInIndexer(obj client.Object) []string {
-	freight := obj.(*Freight) // nolint: forcetypeassert
+	freight := obj.(*kargoapi.Freight) // nolint: forcetypeassert
 	var verifiedIn []string
 	for stage := range freight.Status.VerifiedIn {
 		verifiedIn = append(verifiedIn, stage)
@@ -49,17 +51,17 @@ func verifiedInIndexer(obj client.Object) []string {
 
 func TestGetWarehouse(t *testing.T) {
 	scheme := k8sruntime.NewScheme()
-	require.NoError(t, SchemeBuilder.AddToScheme(scheme))
+	require.NoError(t, kargoapi.SchemeBuilder.AddToScheme(scheme))
 
 	testCases := []struct {
 		name       string
 		client     client.Client
-		assertions func(*testing.T, *Warehouse, error)
+		assertions func(*testing.T, *kargoapi.Warehouse, error)
 	}{
 		{
 			name:   "not found",
 			client: fake.NewClientBuilder().WithScheme(scheme).Build(),
-			assertions: func(t *testing.T, warehouse *Warehouse, err error) {
+			assertions: func(t *testing.T, warehouse *kargoapi.Warehouse, err error) {
 				require.NoError(t, err)
 				require.Nil(t, warehouse)
 			},
@@ -68,14 +70,14 @@ func TestGetWarehouse(t *testing.T) {
 		{
 			name: "found",
 			client: fake.NewClientBuilder().WithScheme(scheme).WithObjects(
-				&Warehouse{
+				&kargoapi.Warehouse{
 					ObjectMeta: metav1.ObjectMeta{
 						Name:      "fake-warehouse",
 						Namespace: "fake-namespace",
 					},
 				},
 			).Build(),
-			assertions: func(t *testing.T, warehouse *Warehouse, err error) {
+			assertions: func(t *testing.T, warehouse *kargoapi.Warehouse, err error) {
 				require.NoError(t, err)
 				require.Equal(t, "fake-warehouse", warehouse.Name)
 				require.Equal(t, "fake-namespace", warehouse.Namespace)
@@ -98,7 +100,7 @@ func TestGetWarehouse(t *testing.T) {
 	}
 }
 
-func TestWarehouse_ListFreight(t *testing.T) {
+func TestListFreightFromWarehouse(t *testing.T) {
 	const testProject = "fake-project"
 	const testWarehouse = "fake-warehouse"
 	const testStage = "fake-stage"
@@ -110,7 +112,7 @@ func TestWarehouse_ListFreight(t *testing.T) {
 		opts        *ListWarehouseFreightOptions
 		objects     []client.Object
 		interceptor interceptor.Funcs
-		assertions  func(*testing.T, []Freight, error)
+		assertions  func(*testing.T, []kargoapi.Freight, error)
 	}{
 		{
 			name: "error listing Freight",
@@ -124,7 +126,7 @@ func TestWarehouse_ListFreight(t *testing.T) {
 					return errors.New("something went wrong")
 				},
 			},
-			assertions: func(t *testing.T, freight []Freight, err error) {
+			assertions: func(t *testing.T, freight []kargoapi.Freight, err error) {
 				require.ErrorContains(t, err, "error listing Freight")
 				require.ErrorContains(t, err, "something went wrong")
 				require.Nil(t, freight)
@@ -133,38 +135,38 @@ func TestWarehouse_ListFreight(t *testing.T) {
 		{
 			name: "success with no options",
 			objects: []client.Object{
-				&Freight{ // This should be returned
+				&kargoapi.Freight{ // This should be returned
 					ObjectMeta: metav1.ObjectMeta{
 						Namespace: testProject,
 						Name:      "fake-freight",
 					},
-					Origin: FreightOrigin{
-						Kind: FreightOriginKindWarehouse,
+					Origin: kargoapi.FreightOrigin{
+						Kind: kargoapi.FreightOriginKindWarehouse,
 						Name: testWarehouse,
 					},
 				},
-				&Freight{ // This should not be returned
+				&kargoapi.Freight{ // This should not be returned
 					ObjectMeta: metav1.ObjectMeta{
 						Namespace: testProject,
 						Name:      "another-fake-freight",
 					},
-					Origin: FreightOrigin{
-						Kind: FreightOriginKindWarehouse,
+					Origin: kargoapi.FreightOrigin{
+						Kind: kargoapi.FreightOriginKindWarehouse,
 						Name: "wrong-warehouse",
 					},
 				},
-				&Freight{ // This should not be returned
+				&kargoapi.Freight{ // This should not be returned
 					ObjectMeta: metav1.ObjectMeta{
 						Namespace: "wrong-project",
 						Name:      "another-fake-freight",
 					},
-					Origin: FreightOrigin{
-						Kind: FreightOriginKindWarehouse,
+					Origin: kargoapi.FreightOrigin{
+						Kind: kargoapi.FreightOriginKindWarehouse,
 						Name: testWarehouse,
 					},
 				},
 			},
-			assertions: func(t *testing.T, freight []Freight, err error) {
+			assertions: func(t *testing.T, freight []kargoapi.Freight, err error) {
 				require.NoError(t, err)
 				require.Len(t, freight, 1)
 				require.Equal(t, testProject, freight[0].Namespace)
@@ -179,81 +181,81 @@ func TestWarehouse_ListFreight(t *testing.T) {
 				VerifiedBefore: &metav1.Time{Time: time.Now().Add(-1 * time.Hour)},
 			},
 			objects: []client.Object{
-				&Freight{ // This should not be returned
+				&kargoapi.Freight{ // This should not be returned
 					ObjectMeta: metav1.ObjectMeta{
 						Namespace: testProject,
 						Name:      "fake-freight-1",
 					},
-					Origin: FreightOrigin{
-						Kind: FreightOriginKindWarehouse,
+					Origin: kargoapi.FreightOrigin{
+						Kind: kargoapi.FreightOriginKindWarehouse,
 						Name: "wrong-warehouse",
 					},
-					Status: FreightStatus{
+					Status: kargoapi.FreightStatus{
 						// Doesn't matter that it's approved for the stage, because this is
 						// the wrong warehouse
-						ApprovedFor: map[string]ApprovedStage{testStage: {}},
+						ApprovedFor: map[string]kargoapi.ApprovedStage{testStage: {}},
 						// Doesn't matter that it's verified upstream, because this is the
 						// wrong warehouse
-						VerifiedIn: map[string]VerifiedStage{testUpstreamStage: {}},
+						VerifiedIn: map[string]kargoapi.VerifiedStage{testUpstreamStage: {}},
 					},
 				},
-				&Freight{ // This should not be returned
+				&kargoapi.Freight{ // This should not be returned
 					ObjectMeta: metav1.ObjectMeta{
 						Namespace: testProject,
 						Name:      "fake-freight-2",
 					},
-					Origin: FreightOrigin{
-						Kind: FreightOriginKindWarehouse,
+					Origin: kargoapi.FreightOrigin{
+						Kind: kargoapi.FreightOriginKindWarehouse,
 						Name: testWarehouse,
 					},
 					// This is not approved or verified in any Stages
 				},
-				&Freight{ // This should be returned
+				&kargoapi.Freight{ // This should be returned
 					ObjectMeta: metav1.ObjectMeta{
 						Namespace: testProject,
 						Name:      "fake-freight-3",
 					},
-					Origin: FreightOrigin{
-						Kind: FreightOriginKindWarehouse,
+					Origin: kargoapi.FreightOrigin{
+						Kind: kargoapi.FreightOriginKindWarehouse,
 						Name: testWarehouse,
 					},
-					Status: FreightStatus{
+					Status: kargoapi.FreightStatus{
 						// This is approved for the Stage
-						ApprovedFor: map[string]ApprovedStage{testStage: {}},
+						ApprovedFor: map[string]kargoapi.ApprovedStage{testStage: {}},
 					},
 				},
-				&Freight{ // This should not be returned
+				&kargoapi.Freight{ // This should not be returned
 					ObjectMeta: metav1.ObjectMeta{
 						Namespace: testProject,
 						Name:      "fake-freight-4",
 					},
-					Origin: FreightOrigin{
-						Kind: FreightOriginKindWarehouse,
+					Origin: kargoapi.FreightOrigin{
+						Kind: kargoapi.FreightOriginKindWarehouse,
 						Name: testWarehouse,
 					},
-					Status: FreightStatus{
+					Status: kargoapi.FreightStatus{
 						// This is verified in the upstream Stage, but the soak time has not
 						// yet elapsed
-						VerifiedIn: map[string]VerifiedStage{
+						VerifiedIn: map[string]kargoapi.VerifiedStage{
 							testUpstreamStage: {
 								VerifiedAt: ptr.To(metav1.Now()),
 							},
 						},
 					},
 				},
-				&Freight{ // This should be returned
+				&kargoapi.Freight{ // This should be returned
 					ObjectMeta: metav1.ObjectMeta{
 						Namespace: testProject,
 						Name:      "fake-freight-5",
 					},
-					Origin: FreightOrigin{
-						Kind: FreightOriginKindWarehouse,
+					Origin: kargoapi.FreightOrigin{
+						Kind: kargoapi.FreightOriginKindWarehouse,
 						Name: testWarehouse,
 					},
-					Status: FreightStatus{
+					Status: kargoapi.FreightStatus{
 						// This is verified in the upstream Stage and the soak time has
 						// elapsed
-						VerifiedIn: map[string]VerifiedStage{
+						VerifiedIn: map[string]kargoapi.VerifiedStage{
 							testUpstreamStage: {
 								VerifiedAt: ptr.To(metav1.NewTime(time.Now().Add(-2 * time.Hour))),
 							},
@@ -261,7 +263,7 @@ func TestWarehouse_ListFreight(t *testing.T) {
 					},
 				},
 			},
-			assertions: func(t *testing.T, freight []Freight, err error) {
+			assertions: func(t *testing.T, freight []kargoapi.Freight, err error) {
 				require.NoError(t, err)
 				require.Len(t, freight, 2)
 				require.Equal(t, testProject, freight[0].Namespace)
@@ -273,46 +275,47 @@ func TestWarehouse_ListFreight(t *testing.T) {
 		{
 			name: "success with AvailabilityStrategy set to FreightAvailabilityStrategyAll",
 			opts: &ListWarehouseFreightOptions{
-				AvailabilityStrategy: FreightAvailabilityStrategyAll,
+				AvailabilityStrategy: kargoapi.FreightAvailabilityStrategyAll,
 				ApprovedFor:          testStage,
 				VerifiedIn:           []string{testUpstreamStage, testUpstreamStage2},
 				VerifiedBefore:       &metav1.Time{Time: time.Now().Add(-1 * time.Hour)},
 			},
 			objects: []client.Object{
-				&Freight{ // This should be returned as its approved for the Stage
+				&kargoapi.Freight{ // This should be returned as its approved for the Stage
 					ObjectMeta: metav1.ObjectMeta{
 						Namespace: testProject,
 						Name:      "fake-freight-1",
 					},
-					Origin: FreightOrigin{
-						Kind: FreightOriginKindWarehouse,
+					Origin: kargoapi.FreightOrigin{
+						Kind: kargoapi.FreightOriginKindWarehouse,
 						Name: testWarehouse,
 					},
-					Status: FreightStatus{
+					Status: kargoapi.FreightStatus{
 						// This is approved for the Stage
-						ApprovedFor: map[string]ApprovedStage{testStage: {}},
+						ApprovedFor: map[string]kargoapi.ApprovedStage{testStage: {}},
 						// This is only verified in a single upstream Stage
-						VerifiedIn: map[string]VerifiedStage{
+						VerifiedIn: map[string]kargoapi.VerifiedStage{
 							testUpstreamStage: {
 								VerifiedAt: ptr.To(metav1.Now()),
 							},
 						},
 					},
 				},
-				&Freight{ // This should be returned because its verified in both upstream Stages and soak time has lapsed
+				&kargoapi.Freight{
+					// This should be returned because its verified in both upstream Stages and soak time has lapsed
 					ObjectMeta: metav1.ObjectMeta{
 						Namespace: testProject,
 						Name:      "fake-freight-2",
 					},
-					Origin: FreightOrigin{
-						Kind: FreightOriginKindWarehouse,
+					Origin: kargoapi.FreightOrigin{
+						Kind: kargoapi.FreightOriginKindWarehouse,
 						Name: testWarehouse,
 					},
-					Status: FreightStatus{
+					Status: kargoapi.FreightStatus{
 						// This is not approved for any Stage
-						ApprovedFor: map[string]ApprovedStage{},
+						ApprovedFor: map[string]kargoapi.ApprovedStage{},
 						// This is verified in all of the upstream Stages and the soak time has lapsed
-						VerifiedIn: map[string]VerifiedStage{
+						VerifiedIn: map[string]kargoapi.VerifiedStage{
 							testUpstreamStage: {
 								VerifiedAt: ptr.To(metav1.NewTime(time.Now().Add(-2 * time.Hour))),
 							},
@@ -322,40 +325,40 @@ func TestWarehouse_ListFreight(t *testing.T) {
 						},
 					},
 				},
-				&Freight{ // This should not be returned because it's not verified in all of the upstream Stages
+				&kargoapi.Freight{ // This should not be returned because it's not verified in all of the upstream Stages
 					ObjectMeta: metav1.ObjectMeta{
 						Namespace: testProject,
 						Name:      "fake-freight-3",
 					},
-					Origin: FreightOrigin{
-						Kind: FreightOriginKindWarehouse,
+					Origin: kargoapi.FreightOrigin{
+						Kind: kargoapi.FreightOriginKindWarehouse,
 						Name: testWarehouse,
 					},
-					Status: FreightStatus{
+					Status: kargoapi.FreightStatus{
 						// This is not approved for any Stage
-						ApprovedFor: map[string]ApprovedStage{},
+						ApprovedFor: map[string]kargoapi.ApprovedStage{},
 						// This is not verified in all of the upstream Stages
-						VerifiedIn: map[string]VerifiedStage{
+						VerifiedIn: map[string]kargoapi.VerifiedStage{
 							testUpstreamStage: {
 								VerifiedAt: ptr.To(metav1.Now()),
 							},
 						},
 					},
 				},
-				&Freight{ // This should not be returned because its not passed the soak time in all Stages
+				&kargoapi.Freight{ // This should not be returned because its not passed the soak time in all Stages
 					ObjectMeta: metav1.ObjectMeta{
 						Namespace: testProject,
 						Name:      "fake-freight-4",
 					},
-					Origin: FreightOrigin{
-						Kind: FreightOriginKindWarehouse,
+					Origin: kargoapi.FreightOrigin{
+						Kind: kargoapi.FreightOriginKindWarehouse,
 						Name: testWarehouse,
 					},
-					Status: FreightStatus{
+					Status: kargoapi.FreightStatus{
 						// This is not approved for any Stage
-						ApprovedFor: map[string]ApprovedStage{},
+						ApprovedFor: map[string]kargoapi.ApprovedStage{},
 						// This is verified in all of the upstream Stages but only passed the soak time of one
-						VerifiedIn: map[string]VerifiedStage{
+						VerifiedIn: map[string]kargoapi.VerifiedStage{
 							testUpstreamStage: {
 								VerifiedAt: ptr.To(metav1.NewTime(time.Now().Add(-2 * time.Hour))),
 							},
@@ -366,7 +369,7 @@ func TestWarehouse_ListFreight(t *testing.T) {
 					},
 				},
 			},
-			assertions: func(t *testing.T, freight []Freight, err error) {
+			assertions: func(t *testing.T, freight []kargoapi.Freight, err error) {
 				require.NoError(t, err)
 				require.Len(t, freight, 2)
 				require.Equal(t, testProject, freight[0].Namespace)
@@ -383,34 +386,36 @@ func TestWarehouse_ListFreight(t *testing.T) {
 				VerifiedIn:           []string{testUpstreamStage, testUpstreamStage2},
 				VerifiedBefore:       &metav1.Time{Time: time.Now().Add(-1 * time.Hour)},
 			},
-			assertions: func(t *testing.T, _ []Freight, err error) {
+			assertions: func(t *testing.T, _ []kargoapi.Freight, err error) {
 				require.ErrorContains(t, err, "unsupported AvailabilityStrategy")
 			},
 		},
 	}
 
 	testScheme := k8sruntime.NewScheme()
-	err := AddToScheme(testScheme)
+	err := kargoapi.AddToScheme(testScheme)
 	require.NoError(t, err)
 
 	for _, testCase := range testCases {
 		c := fake.NewClientBuilder().WithScheme(testScheme).
 			WithScheme(testScheme).
-			WithIndex(&Freight{}, warehouseField, warehouseIndexer).
-			WithIndex(&Freight{}, approvedField, approvedForIndexer).
-			WithIndex(&Freight{}, verifiedInField, verifiedInIndexer).
+			WithIndex(&kargoapi.Freight{}, warehouseField, warehouseIndexer).
+			WithIndex(&kargoapi.Freight{}, approvedField, approvedForIndexer).
+			WithIndex(&kargoapi.Freight{}, verifiedInField, verifiedInIndexer).
 			WithObjects(testCase.objects...).
 			WithInterceptorFuncs(testCase.interceptor).
 			Build()
 
 		t.Run(testCase.name, func(t *testing.T) {
-			warehouse := &Warehouse{
+			warehouse := &kargoapi.Warehouse{
 				ObjectMeta: metav1.ObjectMeta{
 					Namespace: testProject,
 					Name:      "fake-warehouse",
 				},
 			}
-			freight, err := warehouse.ListFreight(context.Background(), c, testCase.opts)
+			freight, err := ListFreightFromWarehouse(
+				context.Background(), c, warehouse, testCase.opts,
+			)
 			testCase.assertions(t, freight, err)
 		})
 	}
