+++ internal/controller/stages/control_flow_stages.go
@@ -16,6 +16,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/source"
 
 	kargoapi "github.com/akuity/kargo/api/v1alpha1"
+	"github.com/akuity/kargo/internal/api"
 	"github.com/akuity/kargo/internal/controller"
 	rollouts "github.com/akuity/kargo/internal/controller/rollouts/api/v1alpha1"
 	"github.com/akuity/kargo/internal/indexer"
@@ -210,7 +211,7 @@ func (r *ControlFlowStageReconciler) Reconcile(ctx context.Context, req ctrl.Req
 	// Ensure the Stage has a finalizer and requeue if it was added.
 	// The reason to requeue is to ensure that a possible deletion of the Stage
 	// directly after the finalizer was added is handled without delay.
-	if ok, err := kargoapi.EnsureFinalizer(ctx, r.client, stage); ok || err != nil {
+	if ok, err := api.EnsureFinalizer(ctx, r.client, stage); ok || err != nil {
 		return ctrl.Result{Requeue: ok}, err
 	}
 
@@ -253,7 +254,7 @@ func (r *ControlFlowStageReconciler) reconcile(
 
 	// Get the available Freight for the Stage.
 	logger.Debug("getting available Freight")
-	freight, err := stage.ListAvailableFreight(ctx, r.client)
+	freight, err := api.ListFreightAvailableToStage(ctx, r.client, stage)
 	if err != nil {
 		newStatus.Message = err.Error()
 		return newStatus, err
@@ -290,7 +291,7 @@ func (r *ControlFlowStageReconciler) initializeStatus(stage *kargoapi.Stage) kar
 	newStatus.Message = ""
 
 	// Record the current refresh token as having been handled.
-	if token, ok := kargoapi.RefreshAnnotationValue(stage.GetAnnotations()); ok {
+	if token, ok := api.RefreshAnnotationValue(stage.GetAnnotations()); ok {
 		newStatus.LastHandledRefresh = token
 	}
 
@@ -359,7 +360,7 @@ func (r *ControlFlowStageReconciler) markFreightVerifiedForStage(
 		r.eventRecorder.AnnotatedEventf(
 			stage,
 			map[string]string{
-				kargoapi.AnnotationKeyEventActor:                  kargoapi.FormatEventControllerActor(r.cfg.Name()),
+				kargoapi.AnnotationKeyEventActor:                  api.FormatEventControllerActor(r.cfg.Name()),
 				kargoapi.AnnotationKeyEventProject:                stage.Namespace,
 				kargoapi.AnnotationKeyEventStageName:              stage.Name,
 				kargoapi.AnnotationKeyEventFreightAlias:           f.Alias,
@@ -415,7 +416,7 @@ func (r *ControlFlowStageReconciler) handleDelete(ctx context.Context, stage *ka
 	}
 
 	// Remove the finalizer from the Stage.
-	if err := kargoapi.RemoveFinalizer(ctx, r.client, stage); err != nil {
+	if err := api.RemoveFinalizer(ctx, r.client, stage); err != nil {
 		return fmt.Errorf("error removing finalizer from Stage: %w", err)
 	}
