+++ internal/server/abort_promotion.go
@@ -7,6 +7,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	kargoapi "github.com/akuity/kargo/api/v1alpha1"
+	"github.com/akuity/kargo/internal/api"
 	svcv1alpha1 "github.com/akuity/kargo/pkg/api/service/v1alpha1"
 )
 
@@ -31,7 +32,7 @@ func (s *server) AbortPromotion(
 		Namespace: project,
 		Name:      name,
 	}
-	if err := kargoapi.AbortPromotion(ctx, s.client, objKey, kargoapi.AbortActionTerminate); err != nil {
+	if err := api.AbortPromotion(ctx, s.client, objKey, kargoapi.AbortActionTerminate); err != nil {
 		return nil, err
 	}
 	return connect.NewResponse(&svcv1alpha1.AbortPromotionResponse{}), nil
