+++ internal/webhook/freight/webhook.go
@@ -20,6 +20,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
 	kargoapi "github.com/akuity/kargo/api/v1alpha1"
+	"github.com/akuity/kargo/internal/api"
 	"github.com/akuity/kargo/internal/git"
 	"github.com/akuity/kargo/internal/helm"
 	"github.com/akuity/kargo/internal/indexer"
@@ -108,7 +109,7 @@ func newWebhook(
 	w.validateProjectFn = libWebhook.ValidateProject
 	w.listFreightFn = kubeClient.List
 	w.listStagesFn = kubeClient.List
-	w.getWarehouseFn = kargoapi.GetWarehouse
+	w.getWarehouseFn = api.GetWarehouse
 	w.validateFreightArtifactsFn = validateFreightArtifacts
 	w.isRequestFromKargoControlplaneFn = libWebhook.IsRequestFromKargoControlplane(cfg.ControlplaneUserRegex)
 	return w
@@ -322,10 +323,10 @@ func (w *webhook) recordFreightApprovedEvent(
 	f *kargoapi.Freight,
 	stageName string,
 ) {
-	actor := kargoapi.FormatEventKubernetesUserActor(req.UserInfo)
+	actor := api.FormatEventKubernetesUserActor(req.UserInfo)
 	w.recorder.AnnotatedEventf(
 		f,
-		kargoapi.NewFreightApprovedEventAnnotations(actor, f, stageName),
+		api.NewFreightApprovedEventAnnotations(actor, f, stageName),
 		corev1.EventTypeNormal,
 		kargoapi.EventReasonFreightApproved,
 		"Freight approved for Stage %q by %q",
